refactor(dashboard): migrate Semantic panel + helpers
Stage 4 PR 4.5 from #28. Largest single-panel migration so far — SemanticPanel + SemanticExcludesCard + ExcludesPreview + ExcludesField + SemanticJobView + SkipBucketsView + toDraft/fromDraft helpers all move to panels/semantic.ts (~530 LoC). dashboard/app.js drops the entire Semantic block.
